requirements.txt
aniso8601==9.0.1
asgiref==3.3.1
astroid==2.4.2
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.4
Deprecated==1.2.10
Django==3.1.5
Flask==2.0.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.0
idna==2.10
isort==5.6.4
itsdangerous==2.0.1
Jinja2==3.0.1
lazy-object-proxy==1.4.3
MarkupSafe==2.0.1
mccabe==0.6.1
prompt-toolkit==1.0.14
pyfiglet==0.8.post1
PyGithub==1.54
Pygments==2.7.3
PyInquirer==1.0.3
PyJWT==2.0.0
pylint==2.6.0
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
requests==2.24.0
six==1.15.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
termcolor==1.1.0
toml==0.10.2
urllib3==1.25.11
wcwidth==0.2.5
Werkzeug==2.0.1
wrapt==1.12.1
